pyproject.toml
[project]
name = "sarumaneyebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "sqlmodel>=0.0.24",
    "textual>=3.5.0",
]
